Validation of equipment section
- Change range of col, row and level

+++ src/components/policy/src/policy/policy_table/table_struct/types.h
@@ -131,9 +131,9 @@ struct Rpcs : CompositeType {
 
 struct InteriorZone: CompositeType {
   public:
-    Integer<uint16_t, 0, 100> col;
-    Integer<uint16_t, 0, 100> row;
-    Integer<uint16_t, 0, 100> level;
+    Integer<uint8_t, 0, 100> col;
+    Integer<uint8_t, 0, 100> row;
+    Integer<uint8_t, 0, 100> level;
     AccessModules auto_allow;
     AccessModules driver_allow;
   public:
